+++ examples/demo/tests/requests/mod.rs
@@ -1,5 +1,8 @@
 mod auth;
+pub mod mylayer;
 mod notes;
+mod ping;
 mod prepare_data;
+mod responses;
 mod upload;
 mod user;