+++ loco-extras/src/lib.rs
@@ -9,4 +9,6 @@
 //!   endpoint.
 //! * `initializer-extra-db` Adding extra DB connection
 //! * `initializer-multi-db` Adding extra DB's connection
+//! * `initializer-normalize-path` Normalize the request path
+//! * `initializer-opentelemetry` For adding opentelemetry tracing
 pub mod initializers;