+++ src/controller/health.rs
@@ -2,11 +2,11 @@
 //! reporting. These routes are commonly used to monitor the health of the
 //! application and its dependencies.
 
-use axum::{extract::State, routing::get};
+use axum::{extract::State, response::Response, routing::get};
 use serde::Serialize;
 
 use super::{format, routes::Routes};
-use crate::{app::AppContext, controller::Json, redis, Result};
+use crate::{app::AppContext, redis, Result};
 
 /// Represents the health status of the application.
 #[derive(Serialize)]
@@ -16,7 +16,7 @@ struct Health {
 
 /// Check the healthiness of the application bt ping to the redis and the DB to
 /// insure that connection
-async fn health(State(ctx): State<AppContext>) -> Result<Json<Health>> {
+async fn health(State(ctx): State<AppContext>) -> Result<Response> {
     let mut is_ok = match ctx.db.ping().await {
         Ok(()) => true,
         Err(error) => {