+++ src/gen/templates/controller.t
@@ -13,12 +13,16 @@ injections:
 ---
 #![allow(clippy::unused_async)]
 use loco_rs::prelude::*;
+use axum::debug_handler;
 
+
+#[debug_handler]
 pub async fn echo(req_body: String) -> String {
     req_body
 }
 
-pub async fn hello(State(_ctx): State<AppContext>) -> Result<String> {
+#[debug_handler]
+pub async fn hello(State(_ctx): State<AppContext>) -> Result<Response> {
     // do something with context (database, etc)
     format::text("hello")
 }