+++ src/gen/templates/scaffold/html/controller.t
@@ -18,6 +18,7 @@ use loco_rs::prelude::*;
 use serde::{Deserialize, Serialize};
 use axum::{extract::Form, response::Redirect};
 use sea_orm::{sea_query::Order, QueryOrder};
+use axum::debug_handler;
 
 use crate::{
     models::_entities::{{file_name | plural}}::{ActiveModel, Column, Entity, Model},
@@ -44,10 +45,11 @@ async fn load_item(ctx: &AppContext, id: i32) -> Result<Model> {
     item.ok_or_else(|| Error::NotFound)
 }
 
+#[debug_handler]
 pub async fn list(
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let item = Entity::find()
         .order_by(Column::Id, Order::Desc)
         .all(&ctx.db)
@@ -55,12 +57,15 @@ pub async fn list(
     views::{{file_name}}::list(&v, &item)
 }
 
+#[debug_handler]
 pub async fn new(
     ViewEngine(v): ViewEngine<TeraView>,
-) -> Result<impl IntoResponse> {
+    State(ctx): State<AppContext>,
+) -> Result<Response> {
     views::{{file_name}}::create(&v)
 }
 
+#[debug_handler]
 pub async fn update(
     Path(id): Path<i32>,
     State(ctx): State<AppContext>,
@@ -73,24 +78,27 @@ pub async fn update(
     Ok(Redirect::to("../{{file_name | plural}}"))
 }
 
+#[debug_handler]
 pub async fn edit(
     Path(id): Path<i32>,
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let item = load_item(&ctx, id).await?;
     views::{{file_name}}::edit(&v, &item)
 }
 
+#[debug_handler]
 pub async fn show(
     Path(id): Path<i32>,
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let item = load_item(&ctx, id).await?;
     views::{{file_name}}::show(&v, &item)
 }
 
+#[debug_handler]
 pub async fn add(
     State(ctx): State<AppContext>,
     Form(params): Form<Params>,