+++ src/gen/templates/scaffold/htmx/controller.t
@@ -17,6 +17,7 @@ injections:
 use loco_rs::prelude::*;
 use serde::{Deserialize, Serialize};
 use sea_orm::{sea_query::Order, QueryOrder};
+use axum::debug_handler;
 
 use crate::{
     models::_entities::{{file_name | plural}}::{ActiveModel, Column, Entity, Model},
@@ -43,10 +44,11 @@ async fn load_item(ctx: &AppContext, id: i32) -> Result<Model> {
     item.ok_or_else(|| Error::NotFound)
 }
 
+#[debug_handler]
 pub async fn list(
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let item = Entity::find()
         .order_by(Column::Id, Order::Desc)
         .all(&ctx.db)
@@ -54,17 +56,20 @@ pub async fn list(
     views::{{file_name}}::list(&v, &item)
 }
 
+#[debug_handler]
 pub async fn new(
     ViewEngine(v): ViewEngine<TeraView>,
-) -> Result<impl IntoResponse> {
+    State(ctx): State<AppContext>,
+) -> Result<Response> {
     views::{{file_name}}::create(v)
 }
 
+#[debug_handler]
 pub async fn update(
     Path(id): Path<i32>,
     State(ctx): State<AppContext>,
     Json(params): Json<Params>,
-) -> Result<Json<Model>> {
+) -> Result<Response> {
     let item = load_item(&ctx, id).await?;
     let mut item = item.into_active_model();
     params.update(&mut item);
@@ -72,29 +77,32 @@ pub async fn update(
     format::json(item)
 }
 
+#[debug_handler]
 pub async fn edit(
     Path(id): Path<i32>,
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let item = load_item(&ctx, id).await?;
     views::{{file_name}}::edit(&v, &item)
 }
 
+#[debug_handler]
 pub async fn show(
     Path(id): Path<i32>,
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let item = load_item(&ctx, id).await?;
     views::{{file_name}}::show(&v, &item)
 }
 
+#[debug_handler]
 pub async fn add(
     ViewEngine(v): ViewEngine<TeraView>,
     State(ctx): State<AppContext>,
     Json(params): Json<Params>,
-) -> Result<impl IntoResponse> {
+) -> Result<Response> {
     let mut item = ActiveModel {
         ..Default::default()
     };