+++ src/lib.rs
@@ -47,7 +47,6 @@ pub mod cli;
 
 pub mod auth;
 pub mod boot;
-pub mod concern;
 pub mod config;
 pub mod controller;
 pub mod environment;
@@ -71,4 +70,4 @@ pub mod tests_cfg;
 pub use validator;
 
 /// Application results options list
-pub type Result<T> = std::result::Result<T, Error>;
+pub type Result<T, E = Error> = std::result::Result<T, E>;