+++ src/prelude.rs
@@ -1,21 +1,29 @@
 pub use async_trait::async_trait;
 pub use axum::{
     extract::{Form, Path, State},
-    response::IntoResponse,
+    response::{IntoResponse, Response},
     routing::{delete, get, post, put},
 };
 pub use axum_extra::extract::cookie;
 pub use chrono::NaiveDateTime as DateTime;
 pub use include_dir::{include_dir, Dir};
 #[cfg(feature = "with-db")]
-pub use sea_orm::{ActiveModelTrait, EntityTrait, IntoActiveModel, ModelTrait, Set};
+pub use sea_orm::{
+    ActiveModelBehavior, ActiveModelTrait, ActiveValue, ColumnTrait, ConnectionTrait,
+    DatabaseConnection, DbErr, EntityTrait, IntoActiveModel, ModelTrait, QueryFilter, Set,
+    TransactionTrait,
+};
 
-#[cfg(any(feature = "auth_jwt", feature = "with-db"))]
+#[cfg(all(feature = "auth_jwt", feature = "with-db"))]
 pub use crate::controller::middleware::auth;
+#[cfg(feature = "with-db")]
+pub use crate::model::{query, Authenticable, ModelError, ModelResult};
 pub use crate::{
     app::{AppContext, Initializer},
     controller::{
-        format, not_found, unauthorized,
+        format,
+        middleware::format::{Format, RespondTo},
+        not_found, unauthorized,
         views::{engines::TeraView, ViewEngine, ViewRenderer},
         Json, Routes,
     },
@@ -23,6 +31,8 @@ pub use crate::{
     mailer,
     mailer::Mailer,
     task::{Task, TaskInfo},
+    validation::{self, Validatable},
+    validator::Validate,
     worker::{self, AppWorker},
     Result,
 };