+++ starters/lightweight-service/src/controllers/home.rs
@@ -1,8 +1,10 @@
+use axum::debug_handler;
 use loco_rs::prelude::*;
 
 use crate::views::home::HomeResponse;
 
-async fn current() -> Result<Json<HomeResponse>> {
+#[debug_handler]
+async fn current() -> Result<Response> {
     format::json(HomeResponse::new("loco"))
 }